+++ braintree/transaction.py
@@ -55,6 +55,7 @@ class Transaction(Resource):
             "amount": "100.00",
             "order_id": "123",
             "channel": "MyShoppingCartProvider",
+            "processing_merchant_category_code": "5411",
             "credit_card": {
                 "number": "5105105105105100",
                 "expiration_date": "05/2011",
@@ -607,7 +608,7 @@ class Transaction(Resource):
                 ]
             },
             "order_id",
-            "payment_method_nonce", "payment_method_token", "product_sku", "purchase_order_number",
+            "payment_method_nonce", "payment_method_token", "processing_merchant_category_code", "product_sku", "purchase_order_number",
             {"paypal_account": ["payee_id", "payee_email", "payer_id", "payment_id"]},
             "recurring",
             {