+++ tests/integration/test_transaction_with_us_bank_account.py
@@ -20,7 +20,6 @@ class TestTransactionWithUsBankAccount(unittest.TestCase):
         self.assertEqual(result.transaction.us_bank_account.account_type, "checking")
         self.assertEqual(result.transaction.us_bank_account.account_holder_name, "Dan Schulman")
         self.assertTrue(re.match(r".*CHASE.*", result.transaction.us_bank_account.bank_name))
-        self.assertEqual(result.transaction.us_bank_account.ach_mandate.text, "cl mandate text")
         self.assertIsInstance(result.transaction.us_bank_account.ach_mandate.accepted_at, datetime)
 
     def test_nonce_transactions_with_vaulted_token(self):
@@ -40,7 +39,6 @@ class TestTransactionWithUsBankAccount(unittest.TestCase):
         self.assertEqual(result.transaction.us_bank_account.account_type, "checking")
         self.assertEqual(result.transaction.us_bank_account.account_holder_name, "Dan Schulman")
         self.assertTrue(re.match(r".*CHASE.*", result.transaction.us_bank_account.bank_name))
-        self.assertEqual(result.transaction.us_bank_account.ach_mandate.text, "cl mandate text")
         self.assertIsInstance(result.transaction.us_bank_account.ach_mandate.accepted_at, datetime)
         token = result.transaction.us_bank_account.token
 
@@ -59,7 +57,6 @@ class TestTransactionWithUsBankAccount(unittest.TestCase):
         self.assertEqual(result.transaction.us_bank_account.account_type, "checking")
         self.assertEqual(result.transaction.us_bank_account.account_holder_name, "Dan Schulman")
         self.assertTrue(re.match(r".*CHASE.*", result.transaction.us_bank_account.bank_name))
-        self.assertEqual(result.transaction.us_bank_account.ach_mandate.text, "cl mandate text")
         self.assertIsInstance(result.transaction.us_bank_account.ach_mandate.accepted_at, datetime)
 
     def test_token_transactions_not_found(self):